requirements.txt
numpy>=1.26
scipy>=1.11
pandas>=2.1
scikit-learn>=1.3
implicit>=0.7.2
faiss-cpu>=1.7.4
tqdm>=4.66
streamlit>=1.32
